fix: binary save and fluorescence selection
